Latest Android updates.

+++ src/com/orsoncharts/android/graphics3d/ViewPoint3D.java
@@ -1,6 +1,6 @@
-/* ============
- * Orson Charts
- * ============
+/* ========================
+ * Orson Charts for Android
+ * ========================
  * 
  * (C)opyright 2013, by Object Refinery Limited.
  * 
@@ -338,7 +338,7 @@ public class ViewPoint3D implements Serializable {
         ViewPoint3D vp = new ViewPoint3D(this.theta, this.phi, this.rho, 
                 calcRollAngle());
         float near = (float) dim3D.getDiagonalLength();
-        float far = (float) near * 40;
+        float far = near * 40f;
         
         World w = new World();
         double ww = dim3D.getWidth();
